# contracts/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plant Doctor RAG API
  version: 0.2.0
  description: |
    식물 주치의 AI MVP용 프론트-백엔드 공유 API 계약 문서입니다.
    
    [업데이트 히스토리]
    * v0.2.0: Supabase pgvector 클라우드 DB 연동 및 RAG 상담 이력 메시지 스키마 정합 반영 완료.
    * v0.1.0: RAG 상담 및 이력 조회 API 엔드포인트 초안 수립.
servers:
  - url: http://localhost:8000
    description: Local backend
security:
  - bearerAuth: []
paths:
...
  /api/v1/chat/sessions:
    get:
      summary: List current user's RAG chat sessions
      parameters:
        - name: plantId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: responseMode
          in: query
          required: false
          schema:
            type: string
            enum: [expert, companion]
          description: expert는 전문가 상담방, companion은 내 식물과 대화하기 상담방만 조회
      responses:
        "200":
          description: List of chat sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
  /api/v1/chat/sessions/{sessionId}/messages:
    get:
      summary: List messages within a chat session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Messages history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
        "404":
          description: Chat session not found or access denied
  /api/v1/chat/plant-care:
    post:
      summary: Run plant care RAG consultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantCareChatRequest"
      responses:
        "200":
          description: RAG answer with citations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantCareChatResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    PlantCareChatRequest:
      type: object
      required: [plantId]
      properties:
        plantId:
          type: string
          format: uuid
        careLogId:
          type: string
          format: uuid
        photoId:
          type: string
          format: uuid
        newSession:
          type: boolean
          default: false
        responseMode:
          type: string
          enum: [expert, companion]
          default: expert
          description: expert는 전문가 상담 말투, companion은 등록한 식물과 대화하는 친근한 말투
        question:
          type: string
          example: 잎 끝이 마르고 아래쪽 잎이 노랗게 변했어요.
    PlantCareChatResponse:
      type: object
      required: [summary, possibleCauses, todayActions, observationChecklist, citations]
      properties:
        summary:
          type: string
        possibleCauses:
          type: array
          items:
            type: string
        todayActions:
          type: array
          items:
            type: string
        observationChecklist:
          type: array
          items:
            type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        safetyNotice:
          type: string
        sessionId:
          type: string
          format: uuid
        messageId:
          type: string
          format: uuid
    Citation:
      type: object
      required: [sourceId, title]
      properties:
        sourceId:
          type: string
        title:
          type: string
        url:
          type: string
        publisher:
          type: string
...
    ChatSession:
      type: object
      required: [id, userId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        plantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    ChatMessage:
      type: object
      required: [id, sessionId, sender, content, createdAt]
      properties:
        id:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        sender:
          type: string
          enum: [user, assistant]
          description: 메시지 전송자 주체 (DB 물리 컬럼명 'role' 과 매핑)
        content:
          type: string
          description: '대화 메시지 본문 텍스트 (DB 물리적으로는 {"text": "내용"} 형태의 JSONB 객체로 보관되나, API 인터페이스 레벨에서는 일반 문자열로 언래핑되어 서빙됩니다.)'
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        createdAt:
          type: string
          format: date-time